docs(wireframe): add wireframe
did two options - option 1 has the most minimal of changes from the current v1 of the app. this allows me to develop fast and build on this in the future potentially. option 2 is a bit more involved, with lots of animations. I'm not sure if I'll be doing this unless I really want to gold-plate this app