feat: on click of a card, Pokemon IDs are regenerated and Pokemons show (#36)
* on initial load of app, when the `<Gameboard />` component renders,
that component's local state (`useState` hook), holding the array of
Pokemon IDs, immediately generates a list of random Pokemon IDs for the
given parameters (we use `getRandomPokemonIds` helper)
* this then triggers the fetching in `useAllPokemons` hook, because
`pokemonIds` are passed in as an argument, and is in the `useEffect`
hook's dependency array
* I've opted to use a hashset for `getRandomPokemonIds` in the `while`
loop because I believe it's more performant than using an array (adding
and searching from a hashset is O(1) time). But, I convert it back into
an array when I `return` because it's easier to manipulate arrays (like
using `.map` in the `useAllPokemons` hook, rather than having to do a
`.forEach` or `for` loop and then returning the array at the end)
* when user clicks on any of the individual cards, the handler passed
down from `<Gameboard />` parent component, as a prop, into `<Card />`
component is triggered. This handler is called `handleMousedownOnCard`.
Yes, I know that right now we're not using the Pokemon ID of the
specific Pokemon card that got clicked, but we'll definitely need this
later on when we implement game logic (need to track this to know if a
Pokemon has been clicked this round)
* this handler then updates the state for Pokemon IDs, which then
triggers a re-render, and updates the arg passed into `useAllPokemons()`
hook, which then triggers `useEffect` within that hook as Pokemon IDs
are in its dependency array, thereby fetching new Pokemon data from
PokeAPI, and then displaying it on the screen