feat(App): set up states for holding high score and current score.
lifted the state all the way up to the App.tsx level. There is some 'prop drilling' happening, when the state is being passed from App, to Header, and then to the Scoreboard component. But, I'll leave it for now unless performance becomes an issue or it just really bothers me subjectively. In that case, I'd opt for something like Context API or an external library like Zustand/Jotai