feat(Gameboard): add local, component state (useState hook) to hold selected Pokemon IDs
makes more sense to just hold IDs (easier to compare too) rather than the whole Pokemon object in that hashset. Hashsets are more efficient at lookup (constant time O(1)) than arrays so opted for this. It's possible that in the future, this should actually be a queue so that I can always see the first pokemon selected this round, or a stack so I can see the last pokemon selected. this is because we might want to hide one or both of those pokemons into the list of pokemons to make it harder (thinking queue is probably better because player will already remember the last one they picked!). anyway, this will come later...