feat(Gameboard): rename PokemonIDs state
clearer and makes distinction between the pokemon IDs for fetching so that the Gameboard can display them (what was changed) vs the pokemon IDs of the pokemons that have been selected (pokemonIdsSelectedThisRound). I know having longer variable names might seem a little not nice (code starts wrapping onto new lines) but if it's clearer, it's better.