feat: handle scores, and losing game (#37)
* when players click on a card, that card's Pokemon ID (unique to each
Pokemon, by business logic of how Pokemons in Pokedex work AFAIK) is
added to a hashset, in the component-level state of Gameboard component.
We use this to check if that Pokemon has been selected before.
* when player clicks on a card that has not been selected before (i.e.
not in the hashset), then we increment the score for the current round.
we update the high score if the current round's score has surpassed it.
* when player clicks on a card that _has_ been selected before (i.e.
included in hashset), then we reset the game. this is handled in the
`resetGame()` method in `Gameboard` component, but essentially: (1)
clears the hashset holding IDs of Pokemons selected, (2) resets the
current round score (but not the high score), and (3) grabs new Pokemons
by refreshing the list of displayed Pokemons IDs

#6 will be done in a different ticket (setting up the modals etc). At
the moment, when the game is lost, we merely show a browser alert, same
as v1:
<img width="662" height="204" alt="image"
src="https://github.com/user-attachments/assets/70b70c1a-6433-4d43-a221-ffdea0ef960f"
/>