feat(useHighScore): create custom hook to get and set high scores to localStorage
needed custom hook to abstract away useEffect syncing of high scores with localStorage such that, if a player has hit a new high score, but then refreshes the page, we still allow them to keep that new high score. otherwise, using a helper and only pushing the high score when the user _loses_ would mean their high score would just restore back to what it was previously. this is a bad UX