feat: persist high scores (#39)
This fixes #4 by using a custom hook (`useHighScore`) to save and fetch
the high score of the user from `localStorage` such that it persists,
without needing a backend.

1. enabling players to get their previous high scores on load of game
2. players' latest high score is saved even if the game hasn't yet ended
-- e.g. if player's previous high score was 1, then the player gets a
high score of 3 while still playing (game has not ended yet) and they
accidentally refresh the page (thereby clearing React state), when the
player sees the game again, they still see 3 (and not 1). This is better
UX.

it is because of point 2 above, that I've opted to use a custom React
hook (to wrap around both `useState` and `useEffect` hooks, the latter
to enable syncing with `localStorage`) rather than just helpers. An
alternative might be to use the helper (something like
`saveHighScoreToLocalStorage()`) after _every click_. **Could try this
if performance becomes an issue** but will keep custom hook for now as
it is, AFAIK, more "React" style