feat(Card): add subtle hover effects
fixes #23. just to give it a bit of interactivity feeling. i took a look at other examples on odin that other people have submitted, and one that was also pokemems used framer motion to animate and a tilt library. i'm trying to not use libraries for the time being and rawdogging this. i think the v1 version (with the yellow background) looks hella tacky so i'm not gonna do that anymore here in v2. might change my mind. who knows.