feat(App): move entire UI down slightly
this is done at the #root div level. however, i didn't just add a justify-content: center, because we don't actually want this app to be in the _true centre_ of the screen, because then the header is too low. what we want, ideally, is the Gameboard to in the centre of players' visions. this transform is just me stabbing in the dark with a magic number i admit :(