feat(Gameboard): collapse grid on narrower screens
but, because there's not a perfect alignment of Card components on the board when it's too narrow (i picked it because some longer-named pokemons like 'weepinbell' started to wrap). opted for a flexbox option with flex-wrap, because with grid, if there are, for example only 2 cards in the bottom row (of 4 columns), i can't force the 2 cards to be in the centre of the row (taking up columns 2 and 3 respectively). because of this pushing the height out, i've also removed the translateY on the #root element of the entire app on narrower screens