feat: make more responsive by adding media queries to Header components (including the Scoreboard).
in Scoreboard, I decided to create a .text class from which the text of the score item and score labels inherit. this allows me to write media queries just on this class, and the score items and label texts just 'composes' this (CSS module). i think this is more DRY.