refactor: add generic Button component, which gets called in PlayAgainButton
not sure if i should still keep playagainbutton as a separate component, when it's only used twice. it feels like an unnecessary layer of abstraction... maybe?? idk, still keeps the codebase DRYer because i don't have to write 'Play Again' twice as the button copy. the purpose of this though was so that i could have a Button component to use in the error modal