feat: add 'retryFetch' method in custom hook and then this gets called at the App level (as the hook needs to have the ids of pokemons passed in). users can then re-call api to get the pokemons.
makes sense for this to be done in the hook because the hook handles all of these behaviours. i also opted for the hook option as opposed to re-generating ids at the App level, updating the displayedPokemonIds state, to trigger re-render, because it isn't, imo, as readable or intuitive as having it in the hook. this also fixes #55.