feat(Modal): add reusable Modal component using <dialog>. Manually test (successfully) that it displays in App.
using <dialog> is more semantically correct anyway that what I was doing before, with a <div> covering everything. I also confirmed that <dialog> prevents me from focusing on the <Card> elements behind when it shows. I've had to use a React ref here, because the way that <dialog>s show and close are done through showModal() and close() methods, which are not part of React's typical props or state updates. useRef provides direct reference to the underlying DOM element (<dialog> here) and call these methods imperatively. useLayoutEffect is used here because state of modal is synchronised. atm, we're using state to control which modal to show, and when to show it (eg isGameOver, isNewHighScore etc). assuming this continues, useLayoutEffect (IIUC) synchronises this state, and also ensures opening of dialog and render occurs in the _same_ render. see stackoverflow re this: https://stackoverflow.com/questions/76567184/html-dialog-element-in-react-js#comment138630148_76567796. the React docs for useLayoutEffect also states that useLayoutEffect is a version of useEffect, but fires _before browser repaints the screen_: https://react.dev/reference/react/useLayoutEffect. i could be wrong and overengineering this, i admit, so i might go back to just using ye good ol' useEffect, since the performance differences would be (assumingly) pretty negligible.