feat: GameOverModal is an instance of Modal. refactor: resetGame method now also updates the state for isGameOver.
by updating state for isGameOver in resetGame, resetGame can be passed as prop alone (without the setter for isGameOver) into GameOverModal as the click handler on the play again button. yes, this might change as I add back the HighScoreModal too and refactor to keep DRY, but for the time being, I think this is a cleaner interface for the GameOverModal. I'm not even sure if we really need a 'play again button' separate component if it's only used in two places, but let's see...