refactor: modals are actually <dialog> elements (#59)
this stemmed from trying to do #57 (fixing issue #24) and issue #58
popping up. instead of writing more JS to try and prevent focus on the
Cards when a modal is shown (like the `<GameOverModal />` for example),
it is more semantic to be using the `<dialog>` JSX elements anyway.

there's just a bit more code requiring the use of `useRef` (and a
`useEffect` / `useLayoutEffect`) but this is abstracted away into the
`<Modal />` reusable UI component (in components/ui).

this fixes #58 

currently the following components are "modals" and now use the reusable
`<Modal />` component:
1. `<ErrorModal />`
2. `<LoadingScreen />`
3. `<GameOverModal />` (shown when user loses, but no new high score)
4. `<HighScoreModal />` (shown when user loses, but new high score has
been set)